Rename 'consume_weight()' to 'consume()'
Unfortunately Rust doesn't support default values for function
parameters, so we decided to come up with two methods to consume tokens:

 * `consume()` that consumes a single token only
 * `consume_weight(N)` that consumes `N` tokens

It's a bit confusing to have two associated functions, and it's a bit
annoying to maintain two associated functions. It shouldn't be too much
challenging for users to always pass how much tokens to consume from the
bucket. So this patch renames `consume_weight()` to `consume()`, and
gets rid of the version that consumes one token only.