Refactor 'consume()' to consume time, not tokens
Currently the token bucket implementation tracks number of tokens in the
bucket, and the last time tokens have been replenished at. Turns out we
can simplify this interface track time only and dynamically count the
amount of tokens we have in the bucket. With this patch we can only have
a single mutable property in the structure.

Co-authored-by: Roman Podoliaka <roman.podoliaka@gmail.com>