Make TokenBucket Sync
Protect `last_replenished_at` with Mutex in order to make sure the token
bucket can be used from multiple threads simultaneously. We should try
to refactor this to use atomic operations instead. That will require
finding an alternative to Instant, though, as it is opaque, and does not
allow us to access the internal u64 counter, which we could increment
atomically.

Co-authored-by: Roman Podoliaka <roman.podoliaka@gmail.com>