Make consume() return a Result instead of a bool
The primary use case for this is so that we can return a custom error,
which would tell the caller for how long they need to wait for the next
attempt to succeed (think Retry-After in HTTP's Too Many Requests status).

We might come up with additional error conditions later (e.g. treat a
rate-limit of 0 as completely blocking a giving key, in which case
retrying does not make sense).

This is much easier now that we removed all the floating point
arithmetic and use a fixed clock in tests.