Allow to block requests by specifying a rate-limit of 0
As an alternative to rate-limiting requests, let's allow to completely
*block* a given entity by specifying the limit of 0. The practical
difference is that a different error should be returned in this case
to emphasize the fact that there is no sense for the caller to retry
requests as they will always be rejected.

While we are at this, add documentation for Error and TokenBucket types.